backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pandas==2.2.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
google-generativeai==0.8.0
